major bug fixes; it was just plain broken.  result has new format.
The calculations were incorrect as the braindead human used the wrong
variable(s) in the wrong spot(s).  The code was cleaned up, and fewer
output piddles generated, as some were not useful for this type of
binning.  For einstance, "minimum" or "maximum" values in each bin are
meaningless as the data may have some underlying order which doesn't
result in a input data set with montonically increasing values.