revamp PDL::PP build setup
* track dependencies with cpanfile

* move from Module::Build to Module::Install

* use standard PDL::PP setup to build, rather than rolling our
  own. Bin1D.pm.PL becomes Bin1D.pm


PDL::PP generates its output in the top directory to be compatible
with XS.  The original .pm.PL layout with Build.PL was an attempt to
get around this.

The latest (as of this writing) ExtUtils::MakeMaker supports the
XSMULTI switch, which allows .xs files to live in the same spot in the
directory hierarchy as the .pm files. However it requires that the XS
files exist when Makefile.PL is run, which isn't the case for PDL::PP
generated XS files, which have to be generated by the PDL that's
installed on the target machine.

Since one day I'd like to get this in PDL, I've decided to switch
to the standard PDL::PP build sequence.  The templating system
which generates the C code is definitely not standard.

I'm much more comfortable with Module::Install than Module::Build, so
the initial experiment with Module::Build was removed.