reset NO_MYMETA, set by pdlpp_stdargs
pdlpp_stdargs sets NO_MYMETA=1. This causes the MYMETA.{json,yml}
files not to be generated during configuration.  However, this breaks
cpanminus, which can't get the extra build and test dependencies.  I'm
not sure why pdlpp_stdargs sets this flag, but this change resets it
to the EU::MM default behavior and makes dependencies work again.